Version 2.3.1
Se agregaron mas filtros a la busqueda de los reportes y se trataron
detalles relacionados a la ultima junta

+++ app/Http/Controllers/ListusersController.php
@@ -166,8 +166,14 @@ class ListusersController extends AppBaseController
 			return redirect(route('listusers.index'));
 		}
 		if($listusers!=Auth::user()){
-			$listusers->delete();
-			Flash::message('Usuario eliminado correctamente.');
+			if($listusers->cabinero()->delete()){
+				$listusers->delete();
+				Flash::message('Usuario eliminado correctamente.');
+			}
+			else{
+				Flash::message('Cabinero no existe');
+			}
+			
 		}else{
 			Flash::message('No puedes eliminarte a ti mismo :).');
 		}
